backend/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in --output-file requirements-dev.txt
aiosqlite==0.21.0
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asteval==1.0.7
    # via vpype
bcrypt==4.3.0
    # via pwdlib
cachetools==6.2.1
    # via vpype
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
    #   vpype
    #   vpype-gcode
coverage==7.11.3
    # via pytest-cov
cryptography==46.0.3
    # via pyjwt
deprecated==1.3.1
    # via limits
dill==0.4.0
    # via multiprocess
distlib==0.4.0
    # via virtualenv
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   fastapi-users
    #   pydantic
execnet==2.1.2
    # via pytest-xdist
fastapi==0.121.1
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-cli==0.0.16
    # via fastapi
fastapi-cloud-cli==0.3.1
    # via fastapi-cli
fastapi-users==15.0.1
    # via
    #   -r requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==7.0.0
    # via fastapi-users
filelock==3.20.0
    # via
    #   torch
    #   virtualenv
fsspec==2025.10.0
    # via torch
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements-dev.in
    #   fastapi
    #   fastapi-cloud-cli
identify==2.6.15
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
limits==5.6.0
    # via slowapi
makefun==1.16.0
    # via fastapi-users
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multiprocess==0.70.18
    # via vpype
mypy==1.18.2
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via mypy
networkx==3.5
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.6
    # via
    #   -r requirements.in
    #   opencv-python
    #   pnoise
    #   scipy
    #   shapely
    #   torchvision
    #   vpype
    #   vpype-gcode
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via
    #   limits
    #   pytest
pathspec==0.12.1
    # via mypy
pillow==12.0.0
    # via
    #   -r requirements.in
    #   pillow-heif
    #   torchvision
pillow-heif==1.1.1
    # via -r requirements.in
platformdirs==4.5.0
    # via virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pnoise==0.2.0
    # via vpype
pre-commit==4.4.0
    # via -r requirements-dev.in
pwdlib==0.2.1
    # via fastapi-users
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via fastapi-users
pyphen==0.15.0
    # via vpype
pytest==9.0.1
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-xdist
pytest-asyncio==1.3.0
    # via -r requirements-dev.in
pytest-cov==7.0.0
    # via -r requirements-dev.in
pytest-xdist==3.8.0
    # via -r requirements-dev.in
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-users
pyyaml==6.0.3
    # via
    #   pre-commit
    #   uvicorn
redis==7.0.1
    # via -r requirements.in
requests==2.32.5
    # via resend
resend==2.19.0
    # via -r requirements.in
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
ruff==0.14.4
    # via -r requirements-dev.in
scipy==1.16.3
    # via vpype
sentry-sdk==2.44.0
    # via fastapi-cloud-cli
setuptools==80.9.0
    # via torch
shapely==2.1.2
    # via vpype
shellingham==1.5.4
    # via typer
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via fastapi-users-db-sqlalchemy
starlette==0.49.3
    # via fastapi
svgelements==1.9.6
    # via vpype
svgwrite==1.4.3
    # via vpype
sympy==1.14.0
    # via torch
tomli==2.3.0
    # via vpype
torch==2.9.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.24.1
    # via -r requirements.in
typer==0.20.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
types-aiofiles==25.1.0.20251011
    # via -r requirements-dev.in
typing-extensions==4.15.0
    # via
    #   aiosqlite
    #   fastapi
    #   limits
    #   mypy
    #   pydantic
    #   pydantic-core
    #   resend
    #   rich-toolkit
    #   sqlalchemy
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.38.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
virtualenv==20.35.4
    # via pre-commit
vpype==1.15.0
    # via
    #   -r requirements.in
    #   vpype-gcode
vpype-gcode==0.13.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   uvicorn
wrapt==2.0.1
    # via deprecated
